pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "silo-smasher"
version = "0.1.0"
description = "Autonomous Root-Cause Investigator — Graph-Augmented Agentic BI that correlates structured data, internal graph context, and external signals to explain why business metrics change."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "airbyte-api==0.53.0",
  "boto3==1.40.67",
  "fastapi>=0.115.0",
  "google-genai==1.65.0",
  "neo4j==5.28.2",
  "openai>=1.0.0",
  "python-dotenv==1.2.1",
  "requests==2.32.5",
  "uvicorn[standard]>=0.32.0",
]

[project.scripts]
build-agent-context = "silo_smasher.cli.build_agent_context:main"
run-diagnostic-orchestrator = "silo_smasher.cli.run_diagnostic_orchestrator:main"
sync-graph-context = "silo_smasher.cli.sync_graph_context:main"
query-graph-rag = "silo_smasher.cli.query_graph_rag:main"

[tool.setuptools.packages.find]
where = ["src"]
